[experimental] A function to patch up JSON RPC params that are destined for the Solana JSON RPC ## Summary
`u64` values, in the Solana JSON RPC, are implemented as JavaScript numbers. This means that any number greater than `Number.MAX_SAFE_INTEGER` (9007199254740991) can't be trusted. This is tracked in https://github.com/solana-labs/solana/issues/30741.

In this PR we create a serializer that converts `bigint` values coming in from `@solana/rpc-core` to JavaScript `number`, as a concession.

## Test Plan

```shell
pnpm test:unit:node
```